Substitute Pathname to get tests to pass in ruby 1.9.3
Moved pathname.rb from ruby 1.9.2 into project in order to continue compatibility to 1.9.3 where
the functionality has been moved into c. In the c version of the file, the unset_const trick
doesn't work because the constant rb_cFile in the c file is never modified. I think this is
actually a bug in how ruby MRI operates, but the fix is a little bit complicated.