Clone symlinks properly.
Previously the File.file? or File.directory? checks would match on symlinks since they resolve the symlink and check the target file. The only time the File.symlink? branch would be entered is when the symlink did not point to a valid file or directory. In that case an arity error would be raised because there were no arguments passed to FileUtils.ln_s.