Fixed premature closing of IO in Pathname#each_line
The previous implementation of Pathname#each_line (see #220) wrapped
both kinds of invocation (block and block-less) in a File.open closure,
the latter resulting in an Enumerator that attempts to operate on a
prematurely closed IO object (resulting in an IOError upon actual
iteration). This fix properly implements the block-less invocation with
a recursing enumerator.

An additional assertion was added to the Pathname unit tests to cover
the case. Existing uses of assert_equal were also corrected to use the
right "expected" and "actual" argument position.

[ Closes #283 ]