Fix rm_f when deleted dirs does not exist
FileUtils.rm_f was implemented as an alias to FileUtils.rm.
This is wrong because an exception is raise if the file does not exist.

This change is similar to an earlier commit to fix the same problem with
FileUtils.rm_rf. (i.e., commit  ab3b1d1e75973f2faadd9d52b3111e6fa1e82c06)