Convert Fixnum to Integer
Because Fixnum has been deprecated and this cleans up a Ruby warning.

In the method has been slightly changed to wrap the values in
Integer(arg, 10) which handles ensuring that the values are integers.
This method will throw TypeError if the values cannot be reasonably
converted to an Integer.

The documentation for Integer(arg, 10) can be found here:
https://ruby-doc.org/core-2.4.0/Kernel.html#method-i-Integer