Add tests for Temfile and IO interaction
While trying to debug a hard to trace error where a Tempfile was created
and then read with IO.foreach, I couldn't explain it. I first started
with a regular minitest reproducer but failed. I then tried to write a
similar test in the original project which was using rspec. There I
could instantly reproduce it. Copying the same test to FakeFS also
worked which means it must be something with the RSpec interaction.

At this point I'm stuck and can't explain why this is but a reproducer
should make it easier for a FakeFS developer to debug.